+++ src/output.rs
@@ -26,11 +26,11 @@ pub enum Output<Datum> {
 #[derive(Clone, PartialEq, Debug, Eq, Deserialize, Serialize)]
 pub struct OutputId {
     tx_hash: String,
-    index: u32,
+    index: u64,
 }
 
 impl OutputId {
-    pub fn new(tx_hash: String, index: u32) -> Self {
+    pub fn new(tx_hash: String, index: u64) -> Self {
         OutputId { tx_hash, index }
     }
 }
@@ -38,7 +38,7 @@ impl OutputId {
 pub type Value = (PolicyId, u64);
 
 impl<Datum> Output<Datum> {
-    pub fn new_wallet(tx_hash: String, index: u32, owner: Address, values: Values) -> Self {
+    pub fn new_wallet(tx_hash: String, index: u64, owner: Address, values: Values) -> Self {
         let id = OutputId::new(tx_hash, index);
         Output::Wallet { id, owner, values }
     }
